add method addDiscount to User class and move methods makePurchase and printTotalAmountOfSpentMoney from different packages to one User package
