+++ src/HomeWork5/UserFunctionalCheck.java
@@ -1,7 +1,8 @@
 package HomeWork5;
 
-import static HomeWork5.CashMethods.makePurchase;
-import static HomeWork5.CashMethods.printTotalAmountOfSpentMoney;
+import static HomeWork5.User.makePurchase;
+import static HomeWork5.User.printTotalAmountOfSpentMoney;
+import static HomeWork5.User.addDiscount;
 
 public class UserFunctionalCheck {
     public static void main(String[] args) {
@@ -12,6 +13,10 @@ public class UserFunctionalCheck {
         firstUser.setPassword("123456");
         firstUser.isActive = true;
         int firstUserSumOfPurchase = 200;
+        int userDiscount = 10;
+        firstUser.setDiscount(10);
+        double sum = firstUser.getAmountSpentMoney();
+        firstUser.setAmountSpentMoney(makePurchase(sum, firstUserSumOfPurchase));
 
         User secondUser = new User();
         secondUser.name = "John Smith";
@@ -25,6 +30,7 @@ public class UserFunctionalCheck {
         System.out.println(firstUser.getPassword());
         System.out.println(firstUser.isActive);
         System.out.println(printTotalAmountOfSpentMoney(makePurchase(firstUserSumOfPurchase, firstUser.getAmountSpentMoney())));
+        addDiscount(userDiscount);
 
         System.out.println("============================");
 
@@ -34,6 +40,6 @@ public class UserFunctionalCheck {
         System.out.println(secondUser.isActive);
         System.out.println(printTotalAmountOfSpentMoney(makePurchase(secondUserSumOfPurchase, secondUser.getAmountSpentMoney())));
 
-
     }
+
 }
